separate deploy function
